Update Toolkit.js
Add in checks for when returnError is false in Toolkit.js.
This is case occurs when a call from deprecated class goes through
Toolkit.js. Instead of worrying about setting the flag on and off
create a new Connection object when same config when returnError is
false within the constructor. Then have proper checks for when to report
the error.