fix: Faulty error thrown by ProgramCall.addParam
When ProgramCall.addParam is passed a DS (2D array) the second parameter is an optional object for parm options.

addParam(data, [,options])

Latest code in master throws an error when passing an DS and no options object to ProgramCall.addParam.

`Error: Specifying the parameter type is required.`

An error should not be thrown if a DS is passed to addParam without providing the option object.