build: Update dep vers and add package-lock.json
A couple packages specified exact versions and those were changed to
allow minor version upgrades.

Based on some research it seems best practice is to commit the
package-lock.json. This is required for use by `npm ci` command, which
is in the default GitHub Actions workflow for npm packages.

Fixes #145