test: Use separate data queue for clear test
When tests are running on multiple systems clearing the data queue
causes race condition with the send/receive tests. This will cause
problems in the CI as Node 8, 10, and 12 kick off tests around the same
time.

Currently, all the tests can share the same send/receive queue because
they all send and expect to receive the same message. So it doesn't
really matter who the actual sender or receiver are. If we ever add
other tests that expect to send/receive different messages we'll hvae to
do something else to ensure they have a unique data queue.