test: Fix race conditions in user space tests
Unlike the data queue tests, the user space tests expect to create and
delete the user space being tested with as well as modify the data in
the user space and test that it was read correctly. This means there is
no way to run these tests using the same userspace object simultaneously
with other test runners without causing race conditions.

Instead, we generate a random name and use it. The names must start with
a letter and then can be letters or numbers. A ten character name (max)
gives a 1 in 2,640,558,873,378,816 chance of a collision, which is
unlikely enough for me. :)