fix: Move iConn compatibility code to iConn itself
In v1, iConn is now a wrapper around Connection, however the
compatibility code was added to the Connection constructor instead of
having iConn do that, which means that Connection still had the same
signature as iConn and users could still use it the same way (though it
was not going to be documented as such). Instead, iConn should handle
its own compatibilty layer itself so that when the deprecated function
is removed, the compatibility layer is removed as well and it prevents
users from using Connection in a way that is not supported.